re organize waiver1915b templates, update tests and snapshots accordingly
