+++ lib/libs/email/preview/InitialSubmissions/CMS/InitialSubmissionCMS.test.tsx
@@ -5,8 +5,13 @@ import AppKCMSEmailPreview from "./__snapshots__/AppK";
 import ChipSpaCMSEmailPreview from "./CHIP_SPA";
 import Medicaid_SPA from "./Medicaid_SPA";
 import TempExtCMSPreview from "./Temp_Extension";
-import * as WaiverCapitated from "./Waiver_Capitated";
-import * as WaiverContracting from "./Waiver_Contracting";
+import CMSWaiver1915bCapitatedInitial from "./Waiver1915bCapitated/Initial";
+import CMSWaiver1915bCapitatedRenewal from "./Waiver1915bCapitated/Renewal";
+import CMSWaiver1915bCapitatedAmendment from "./Waiver1915bCapitated/Amendment";
+import CMSWaiver1915bContractingInitial from "./Waiver1915bContracting/Initial";
+import CMSWaiver1915bContractingRenewal from "./Waiver1915bContracting/Renewal";
+import CMSWaiver1915bContractingAmendment from "./Waiver1915bContracting/Amendment";
+
 describe("Initial Submission CMS Email Snapshot Test", () => {
   it("renders a AppkCMSEmail Preview Template", () => {
     const template = render(<AppKCMSEmailPreview />);
@@ -29,32 +34,32 @@ describe("Initial Submission CMS Email Snapshot Test", () => {
     expect(template).toMatchSnapshot();
   });
   it("renders a Initial Waiver Capitated Preview Template", () => {
-    const template = render(<WaiverCapitated.Waiver1915bCMSCapitatedInitialEmailPreview />);
+    const template = render(<CMSWaiver1915bCapitatedInitial />);
 
     expect(template).toMatchSnapshot();
   });
   it("renders a Renewal Waiver Capitated Preview Template", () => {
-    const template = render(<WaiverCapitated.Waiver1915bCMSCapitatedRenewalEmailPreview />);
+    const template = render(<CMSWaiver1915bCapitatedRenewal />);
 
     expect(template).toMatchSnapshot();
   });
   it("renders a Amendment Waiver Capitated Preview Template", () => {
-    const template = render(<WaiverCapitated.Waiver1915bCMSCapitatedAmendmentEmailPreview />);
-
-    expect(template).toMatchSnapshot();
-  });
-  it("renders a Amendment Waiver Contracting Preview Template", () => {
-    const template = render(<WaiverContracting.Waiver1915bCMSContractingAmendmentEmailPreview />);
-
-    expect(template).toMatchSnapshot();
-  });
-  it("renders a Renewal Waiver Contracting Preview Template", () => {
-    const template = render(<WaiverContracting.Waiver1915bCMSContractingRenewalEmailPreview />);
+    const template = render(<CMSWaiver1915bCapitatedAmendment />);
 
     expect(template).toMatchSnapshot();
   });
   it("renders a Initial Waiver Contracting Preview Template", () => {
-    const template = render(<WaiverContracting.Waiver1915bCMSContractingInitialEmailPreview />);
+    const template = render(<CMSWaiver1915bContractingInitial />);
+
+    expect(template).toMatchSnapshot();
+  });
+  it("renders a Renewal Waiver Contracting Preview Template", () => {
+    const template = render(<CMSWaiver1915bContractingRenewal />);
+
+    expect(template).toMatchSnapshot();
+  });
+  it("renders a Amendment Waiver Contracting Preview Template", () => {
+    const template = render(<CMSWaiver1915bContractingAmendment />);
 
     expect(template).toMatchSnapshot();
   });
